pyproject.toml
[project]
name = "ntp-calc-images-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "odfpy>=1.4.1",
    "psycopg2-binary>=2.9.10",
]
